a big refactor: * move all model (mdl) transformation related code into "common" crate * make "MAX_MVP" to usize * remove all model entity updating related coed because im still fixing it * make "mvp_buffer" module solely responsible for creating mvp * reduce world mvp buffer from two buffers to one buffer, this makes shader looks simpler * first draft of "world dynamic buffer" where we can load anything during runtime here *

+++ kdr/src/app/constants.rs
@@ -2,7 +2,7 @@ pub const DEFAULT_NOCLIP_SPEED: f32 = 1000.;
 pub const CAM_TURN: f32 = 150.; // degrees
 pub const DEFAULT_SENSITIVITY: f32 = 0.06; // degree angles per mouse unit moved
 
-pub const MAX_MVP: u32 = 1024;
+pub const MAX_MVP: usize = 1024;
 
 pub const WINDOW_MINIMUM_WIDTH: u32 = 640;
 pub const WINDOW_MINIMUM_HEIGHT: u32 = 480;
